Separate color storage from flags (such as visited)

+++ hextile.js
@@ -10,7 +10,7 @@ function OddQHexTile(origin, width, height) {
     this.origin = origin.toOddQOffset();
     this.width = width;
     this.height = height;
-    this.data = new Uint8Array(this.width * this.height);
+    this.data = new Uint16Array(this.width * this.height);
 }
 
 OddQHexTile.prototype.boundingBox = function boundingBox() {
